requirements/requirements.txt
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.0
tensorflow>=2.4.0
joblib>=1.0.0
paho-mqtt>=1.5.1
python-dotenv>=0.15.0
matplotlib>=3.3.3
seaborn>=0.11.1
